part3 update
Completed part three. Ruby is capable of turning certain string formats back into Hash's. Unfortunately the dictionary did not come in the right format. Luckily, ruby contains a substitution method for strings. Substituted the colons for arrows to follow the correct format. Turned the dictionary into a hash. Then, I flattened the hash by one level, this allows for the string (no matter what it is, because the key could be anything) to be on the same level as the array (can't get the key, so I can't get the value. But, I can put everything on the same level and access them from one flat array). Then I used the array's built in index method (although I could have easily used array.each) to find the index.