Part 4/5
Was concerned that some strings may contain "=>" which could create potentially create errors given specific string patterns. So, I switched to using JSON's parse method. Then I used Ruby's "start_with" string method to make sure that it's only matching with strings that being with the prefix (using 'contains' could cause potential errors if prefix is located in the middle of a string). Then, I just use an if statement that only adds to the array if it does not have the prefix.