Part 5
imported Ruby's time library. This allowed me to convert the time given to me by the API into a time object which has an add function that lets me simply add seconds. Then, to get the string back to original format (currently YYYY-MM-DD HH:MM:SS UTC), I split the string up by spaces which gave me an array for the date the time and "UTC". Then I just added the date and Time together with a "T" and added a "Z" back to the end